+++ examples/chat/assets/bundle.ts
@@ -1,0 +1,6 @@
+// Copyright 2022 Connor Logan. All rights reserved. MIT License.
+// This module is browser-only.
+
+export { initLanding } from "./index.ts";
+export { initChat } from "../pages/chat.ts";
+export { initLogin } from "../pages/login.ts";