+++ serial.ts
@@ -857,11 +857,18 @@ export function packResponse(
   init?: PackResponseInit,
 ): Response {
   // If it's already a Response, just append the headers and forward it. Ignore
-  // the init status and statusText
+  // the init status and statusText. If appending the headers fails, just return
+  // the response. (This can happen if the response is a redirect response, for
+  // example.)  
+  // TODO: Test the above caveat
   if (body instanceof Response) {
-    const mergeHeaders = new Headers(init?.headers);
-    for (const [k, v] of mergeHeaders.entries()) {
-      body.headers.append(k, v);
+    try {
+      const mergeHeaders = new Headers(init?.headers);
+      for (const [k, v] of mergeHeaders.entries()) {
+        body.headers.append(k, v);
+      }
+    } catch {
+      // continue;
     }
     return body;
   }